clean

+++ code/parachain/frame/composable-traits/src/xcm/memo.rs
@@ -15,7 +15,7 @@ use crate::prelude::*;
 	Debug,
 )]
 pub struct ChainInfo {
-	pub chain_id: u128,
+	pub chain_id: u32,
 	pub order: u8,
 	pub channel_id: u64,        //for packet or memo
 	pub timestamp: Option<u64>, //for packet
